[model]: fix the previous button
The index of the next song was not updated, when we change the song. As a result, each song, that's played by pressing the previous button, will play again, if we press the next button. What this PR does is, sets audio index, to the previous song of the current audio. Then updates the index of the next song.